Move VectorPlus to Functional
